RAPPO/requirements.txt
absl-py==0.12.0
appdirs==1.4.4
astunparse==1.6.3
attrs==21.2.0
auto-LiRPA==0.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudpickle==1.6.0
cox==9
cycler==0.10.0
Cython==0.29.23
debtcollector==2.2.0
dill==0.3.3
et-xmlfile==1.1.0
fasteners==0.16
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.14
glfw==2.1.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gym==0.17.3
h5py==2.10.0
idna==2.10
imageio==2.9.0
iniconfig==1.1.1
iso8601==0.1.14
Jinja2==3.0.3
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mujoco-py==2.0.2.13
netaddr==0.8.0
netifaces==0.10.9
numexpr==2.7.3
numpy==1.20.3
oauthlib==3.1.0
openpyxl==3.0.9
opt-einsum==3.3.0
oslo.concurrency==4.4.0
oslo.config==8.6.0
oslo.i18n==5.0.1
oslo.utils==4.8.0
packaging==20.9
pandas==1.3.0
pbr==5.6.0
Pillow==7.2.0
pkg_resources==0.0.0
pluggy==0.13.1
protobuf==3.16.0
psutil==5.8.0
py==1.10.0
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
six==1.15.0
smmap==4.0.0
stevedore==3.3.0
tables==3.6.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.7.1+cu110
torchaudio==0.7.2
torchdiffeq==0.2.3
torchfile==0.1.0
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
visdom==0.1.8.9
websocket-client==0.59.0
Werkzeug==2.0.0
wrapt==1.12.1
xmltodict==0.12.0
